pyproject.toml
[tool.poetry]
name = "buck-ai-guide-cse-faculties"
version = "0.1.0"
description = "Streamit RAG app utilizing LangChain, Openai, and Pinecone to recommend OSU CSE Department faculties to users."
authors = ["Keming He <keminghe.career@gmail.com>", "Noah Olsen <noah24olsen+hackohio[at]gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.12.*"
python-dotenv = "^1.0.1"
langchain = "^0.3.4"
langchain-openai = "^0.2.3"
streamlit = "^1.39.0"
pydantic = "^2.9.2"
pinecone = "^5.3.1"
langchain-pinecone = "^0.2.0"
pinecone-notebooks = "^0.1.1"

[tool.poetry.group.dev.dependencies]
notebook = "^7.2.2"
ipywidgets = "^8.1.5"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
scraperapi-sdk = "^1.5.3"
lxml = "^5.3.0"
pandas = "^2.2.3"
plotly = "^5.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
